flask/requirements.txt
absl-py==1.2.0
aio-pika==8.2.4
aiofiles==22.1.0
aiogram==2.19
aiohttp==3.8.3
aiormq==6.4.2
aiosignal==1.2.0
aniso8601==9.0.1
APScheduler==3.9.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bidict==0.22.0
blis==0.7.9
boto3==1.26.4
botocore==1.29.4
CacheControl==0.12.11
cachelib==0.9.0
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.1.0
confection==0.0.3
cryptography==38.0.3
cycler==0.11.0
cymem==2.0.7
dask==2022.2.0
Deprecated==1.2.13
dnspython==1.16.0
docopt==0.6.2
en-core-web-md==3.4.1
en-core-web-sm==3.4.1
fbmessenger==6.0.0
fire==0.4.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-Session==0.4.0
flask-sock==0.5.2
Flask-SocketIO==5.3.1
flatbuffers==22.10.26
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gast==0.5.3
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonschema==4.16.0
kafka-python==2.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.2.0rc1
mattermostwrapper==2.2
msgpack==1.0.4
multidict==5.2.0
murmurhash==1.0.9
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==3.2.1
partd==1.3.0
pathy==0.6.2
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
preshed==3.0.8
prompt-toolkit==3.0.28
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.3.0
rasa-sdk==3.3.0
redis==4.3.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rocketchat-API==1.26.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.9.10
simple-websocket==0.8.1
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
smart-open==5.2.1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.43
srsly==2.4.5
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.3
tensorflow-addons==0.17.1
tensorflow-estimator==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-text==2.8.2
termcolor==2.1.0
terminaltables==3.1.10
thinc==8.1.5
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
twilio==7.14.2
typeguard==2.13.3
typer==0.4.2
typing-extensions==4.4.0
typing-utils==0.1.0
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvloop==0.17.0
wasabi==0.10.1
wcwidth==0.2.5
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
zipp==3.10.0
